// client/src/store/action/auth/action.ts
export enum LOGIN_ACTION {
    LOGIN_PENDING = "LOGIN_PENDING",
    LOGIN_SUCCESS = "LOGIN_SUCCESS",
    LOGIN_FAILURE = "LOGIN_FAILURE",
}

export enum REGISTER_ACTION {
    REGISTER_PENDING = "REGISTER_PENDING",
    REGISTER_SUCCESS = "REGISTER_SUCCESS",
    REGISTER_FAILURE = "REGISTER_FAILURE",
}

export enum AUTH_ME_ACTION {
    AUTH_ME_PENDING = "AUTH_ME_PENDING",
    AUTH_ME_SUCCESS = "AUTH_ME_SUCCESS",
    AUTH_ME_FAILURE = "AUTH_ME_FAILURE",
}

export enum LOGOUT_ACTION {
    LOGOUT_PENDING = "LOGOUT_PENDING",
    LOGOUT_SUCCESS = "LOGOUT_SUCCESS",
    LOGOUT_FAILURE = "LOGOUT_FAILURE",
}

export enum REFRESH_TOKEN_ACTION {
    REFRESH_TOKEN_PENDING = "REFRESH_TOKEN_PENDING",
    REFRESH_TOKEN_SUCCESS = "REFRESH_TOKEN_SUCCESS",
    REFRESH_TOKEN_FAILURE = "REFRESH_TOKEN_FAILURE",
}

export enum GET_ALL_USER_ACTION {
    GET_ALL_USER_PENDING = "GET_ALL_USER_PENDING",
    GET_ALL_USER_SUCCESS = "GET_ALL_USER_SUCCESS",
    GET_ALL_USER_FAILURE = "GET_ALL_USER_FAILURE",
}